% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_dataset.R
\name{generate_dataset}
\alias{generate_dataset}
\title{Generate a crafted dataset of spurious importance}
\usage{
generate_dataset(n = 500)
}
\arguments{
\item{n}{the number of rows}
}
\description{
Generate a crafted dataset as an example of important predictors with spuriously low importance.
}
\examples{
ds <- generate_dataset(1000)

}
