% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_dataset.R
\name{generate_dataset}
\alias{generate_dataset}
\title{Generate a crafted dataset of spurious importance}
\usage{
generate_dataset(n = 500, a = 1, b = 0.2, c = 0.02)
}
\arguments{
\item{n}{the number of rows}

\item{a}{the regression coefficient of \eqn{v}}

\item{b}{the regression coefficient of \eqn{x}}

\item{c}{the regression coefficient of \eqn{y}}
}
\description{
Generate a crafted dataset as an example of important predictors with spuriously low importance.
There are five variables \eqn{v, w, x, y,} and \eqn{z} for predictors,
while two other variables \eqn{r} and \eqn{s} are for responses.
Every predictor variable, except \eqn{w}, is generated by following an independent
standard normal distribution.
On the other hand, \eqn{w} is defined by a simple linear equation of \eqn{v}:
\deqn{w = 1 - v.}
\eqn{r}'s true model is a multiple linear regression of the following equation:
\deqn{r = a v + b x + c y.}{r = a * v + b * x + c * y.}
By default all of the regression coefficients \eqn{a, b}, and \eqn{c} are positive;
\eqn{a} is the largest among them;
\eqn{c} is the smallest.
A binary variable \eqn{s} forms a factor with levels \dQuote{A} and \dQuote{B}.
It takes value "A" if \eqn{s} is of the lower ranks, \dQuote{B} otherwise.
}
\examples{
ds <- generate_dataset(1000)

}
